pyproject.toml
[project]
name = "business-case-builder"
version = "0.1.0"
description = "GitHub Copilot agent that builds business cases for features"
requires-python = ">=3.12"
dependencies = [
    "python-docx>=1.1.0",
    "matplotlib>=3.8.0",
    "python-pptx>=1.0.0",
    "azure-devops>=7.1.0b4",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
